tests/head-tag: Fix `id` conflict
The `id='test-comp'` has priority over the the `attrs.id` that we also pass in. This leads to the `assert.equal()` failing for the `id`, because it is `test-comp`, not `the-id`. With this change we remove the `id='test-comp'` and just run the assertions on the `<meta>` tag without an element ID selector. This should work fine, as long as we use `this.element.querySelector()` instead of `document.querySelector()`.