Fix CI builds (#112)
* tests/head-tag: Fix `id` conflict

The `id='test-comp'` has priority over the the `attrs.id` that we also pass in. This leads to the `assert.equal()` failing for the `id`, because it is `test-comp`, not `the-id`. With this change we remove the `id='test-comp'` and just run the assertions on the `<meta>` tag without an element ID selector. This should work fine, as long as we use `this.element.querySelector()` instead of `document.querySelector()`.

* Relax `ember-cli-babel` dependency constraint

ember-cli-babel v7.17.x is incompatible with Ember v4 and there is no reason for us to restrict it to the v7.17.x minor versions.

* mixins/route-meta: Extract `_collectHeadTags()` method

This fixes the sinon mocking in the corresponding test, since `run.next()` can't be mocked that easily anymore within recent versions of Ember

* CI: Use `yarn` instead of `npm`

* CI: Use `actions/setup-node` instead of `volta-cli/action`

This project does not appear to use volta at all, and it fails to find `yarn` this way

* CI: Remove broken `publish` job

There is no `semantic-release` script in this project, so this can't work at all.

* CI: Remove scenarios from `ember-try-scenario` matrix that do not exist

* CI: Disable Ember 4.x scenarios

Fixing compatibility with Ember 4.x appears to be a slightly larger task. We should get CI working again first, before we continue with that.